- Create ultils folder inside test folder - Move common.js to ultils folder above - Change all reference of common.js
