+++ test/module/record/addComment.test.js
@@ -4,7 +4,7 @@
  * test record module
  */
 const nock = require('nock');
-const common = require('../../common');
+const common = require('../../utils/common');
 
 const KintoneAPIException = require('../../../src/exception/KintoneAPIException');
 const Connection = require('../../../src/connection/Connection');
