+++ test/module/record/deleteComment.test.js
@@ -5,7 +5,7 @@
  */
 const nock = require('nock');
 
-const common = require('../../common');
+const common = require('../../utils/common');
 
 const KintoneAPIException = require('../../../src/exception/KintoneAPIException');
 const Connection = require('../../../src/connection/Connection');
