+++ test/module/record/getRecords.test.js
@@ -4,7 +4,7 @@
  */
 
 const nock = require('nock');
-const common = require('../../common');
+const common = require('../../utils/common');
 const Connection = require('../../../src/connection/Connection');
 const Auth = require('../../../src/authentication/Auth');
 const Record = require('../../../src/module/record/Record');
